test(e2e): Align wrapper and auto-instrumentation scenarios and cover Anthropic, Google GenAI, and Claude Agent SDK
