+++ e2e/helpers/ai-sdk.ts
@@ -12,12 +12,18 @@ export interface WrapAISDKScenario {
 }
 
 export interface AISDKAutoHookScenario {
-  agentClassExport: "Experimental_Agent" | "ToolLoopAgent";
-  agentSpanName: "Agent" | "ToolLoopAgent";
+  agentClassExport?: "Experimental_Agent" | "ToolLoopAgent";
+  agentSpanName?: "Agent" | "ToolLoopAgent";
+  dependencyName: string;
   entry: string;
+  maxTokensKey: "maxOutputTokens" | "maxTokens";
+  openaiDependencyName: string;
+  openaiModuleName: string;
+  packageName: string;
   supportsGenerateObject: boolean;
   supportsStreamObject: boolean;
   supportsToolExecution: boolean;
+  toolSchemaKey: "inputSchema" | "parameters";
   version: string;
 }
 
@@ -26,35 +32,59 @@ export const AI_SDK_SCENARIO_TIMEOUT_MS = 120_000;
 const AI_SDK_SCENARIO_SPECS = [
   {
     dependencyName: "ai-sdk-v3",
+    entryMjs: "scenario.ai-sdk-v3.mjs",
     entry: "scenario.ai-sdk-v3.ts",
+    maxTokensKey: "maxTokens",
+    openaiDependencyName: "ai-sdk-openai-v3",
+    openaiModuleName: "ai-sdk-openai-v3",
+    packageName: "ai-sdk-v3",
     supportsGenerateObject: true,
     supportsStreamObject: true,
     supportsToolExecution: false,
+    toolSchemaKey: "parameters",
   },
   {
     dependencyName: "ai-sdk-v4",
+    entryMjs: "scenario.ai-sdk-v4.mjs",
     entry: "scenario.ai-sdk-v4.ts",
+    maxTokensKey: "maxTokens",
+    openaiDependencyName: "ai-sdk-openai-v4",
+    openaiModuleName: "ai-sdk-openai-v4",
+    packageName: "ai-sdk-v4",
     supportsGenerateObject: true,
     supportsStreamObject: true,
     supportsToolExecution: false,
+    toolSchemaKey: "parameters",
   },
   {
     agentClassExport: "Experimental_Agent",
     agentSpanName: "Agent",
     dependencyName: "ai-sdk-v5",
+    entryMjs: "scenario.ai-sdk-v5.mjs",
     entry: "scenario.ai-sdk-v5.ts",
+    maxTokensKey: "maxOutputTokens",
+    openaiDependencyName: "ai-sdk-openai-v5",
+    openaiModuleName: "ai-sdk-openai-v5",
+    packageName: "ai-sdk-v5",
     supportsGenerateObject: true,
     supportsStreamObject: true,
     supportsToolExecution: true,
+    toolSchemaKey: "inputSchema",
   },
   {
     agentClassExport: "ToolLoopAgent",
     agentSpanName: "ToolLoopAgent",
     dependencyName: "ai-sdk-v6",
+    entryMjs: "scenario.ai-sdk-v6.mjs",
     entry: "scenario.ai-sdk-v6.ts",
+    maxTokensKey: "maxOutputTokens",
+    openaiDependencyName: "ai-sdk-openai-v6",
+    openaiModuleName: "ai-sdk-openai-v6",
+    packageName: "ai-sdk-v6",
     supportsGenerateObject: true,
     supportsStreamObject: true,
     supportsToolExecution: true,
+    toolSchemaKey: "inputSchema",
   },
 ] as const;
 
@@ -72,23 +102,27 @@ export async function getWrapAISDKScenarios(
   );
 }
 
-export const AI_SDK_AUTO_HOOK_SCENARIOS: AISDKAutoHookScenario[] = [
-  {
-    agentClassExport: "Experimental_Agent",
-    agentSpanName: "Agent",
-    entry: "scenario.ai-sdk-v5.mjs",
-    supportsGenerateObject: true,
-    supportsStreamObject: true,
-    supportsToolExecution: true,
-    version: "5.0.82",
-  },
-  {
-    agentClassExport: "ToolLoopAgent",
-    agentSpanName: "ToolLoopAgent",
-    entry: "scenario.ai-sdk-v6.mjs",
-    supportsGenerateObject: true,
-    supportsStreamObject: true,
-    supportsToolExecution: true,
-    version: "6.0.1",
-  },
-];
+export async function getAISDKAutoHookScenarios(
+  scenarioDir: string,
+): Promise<AISDKAutoHookScenario[]> {
+  return await Promise.all(
+    AI_SDK_SCENARIO_SPECS.map(async (scenario) => ({
+      agentClassExport: scenario.agentClassExport,
+      agentSpanName: scenario.agentSpanName,
+      dependencyName: scenario.dependencyName,
+      entry: scenario.entryMjs,
+      maxTokensKey: scenario.maxTokensKey,
+      openaiDependencyName: scenario.openaiDependencyName,
+      openaiModuleName: scenario.openaiModuleName,
+      packageName: scenario.packageName,
+      supportsGenerateObject: scenario.supportsGenerateObject,
+      supportsStreamObject: scenario.supportsStreamObject,
+      supportsToolExecution: scenario.supportsToolExecution,
+      toolSchemaKey: scenario.toolSchemaKey,
+      version: await readInstalledPackageVersion(
+        scenarioDir,
+        scenario.dependencyName,
+      ),
+    })),
+  );
+}
