+++ e2e/helpers/mock-claude-agent-sdk/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@anthropic-ai/claude-agent-sdk",
+  "private": true,
+  "version": "0.1.0",
+  "type": "module",
+  "exports": {
+    ".": "./sdk.mjs"
+  }
+}
