+++ e2e/scenarios/ai-sdk-auto-instrumentation-node-hook/scenario.ai-sdk-v5.mjs
@@ -1,10 +1,17 @@
 import { openai } from "ai-sdk-openai-v5";
 import * as ai from "ai-sdk-v5";
+import { getInstalledPackageVersion } from "../../helpers/provider-runtime.mjs";
 import { runAISDKAutoInstrumentationNodeHookOrExit } from "./scenario.impl.mjs";
 
-runAISDKAutoInstrumentationNodeHookOrExit(
+runAISDKAutoInstrumentationNodeHookOrExit({
+  agentClassExport: "Experimental_Agent",
   ai,
+  agentSpanName: "Agent",
+  maxTokensKey: "maxOutputTokens",
   openai,
-  "5.0.82",
-  "Experimental_Agent",
-);
+  sdkVersion: await getInstalledPackageVersion(import.meta.url, "ai-sdk-v5"),
+  supportsGenerateObject: true,
+  supportsStreamObject: true,
+  supportsToolExecution: true,
+  toolSchemaKey: "inputSchema",
+});
