+++ e2e/scenarios/ai-sdk-auto-instrumentation-node-hook/scenario.ai-sdk-v6.mjs
@@ -1,5 +1,17 @@
 import { openai } from "ai-sdk-openai-v6";
 import * as ai from "ai-sdk-v6";
+import { getInstalledPackageVersion } from "../../helpers/provider-runtime.mjs";
 import { runAISDKAutoInstrumentationNodeHookOrExit } from "./scenario.impl.mjs";
 
-runAISDKAutoInstrumentationNodeHookOrExit(ai, openai, "6.0.1", "ToolLoopAgent");
+runAISDKAutoInstrumentationNodeHookOrExit({
+  agentClassExport: "ToolLoopAgent",
+  agentSpanName: "ToolLoopAgent",
+  ai,
+  maxTokensKey: "maxOutputTokens",
+  openai,
+  sdkVersion: await getInstalledPackageVersion(import.meta.url, "ai-sdk-v6"),
+  supportsGenerateObject: true,
+  supportsStreamObject: true,
+  supportsToolExecution: true,
+  toolSchemaKey: "inputSchema",
+});
