+++ e2e/scenarios/ai-sdk-auto-instrumentation-node-hook/scenario.test.ts
@@ -1,22 +1,21 @@
-import { beforeAll, expect, test } from "vitest";
+import { expect, test } from "vitest";
 import {
-  AI_SDK_AUTO_HOOK_SCENARIOS,
   AI_SDK_SCENARIO_TIMEOUT_MS,
+  getAISDKAutoHookScenarios,
 } from "../../helpers/ai-sdk";
 import { assertAISDKTraceContract } from "../../helpers/ai-sdk-trace-contract";
 import {
-  installScenarioDependencies,
+  prepareScenarioDir,
   resolveScenarioDir,
   withScenarioHarness,
 } from "../../helpers/scenario-harness";
 
-const scenarioDir = resolveScenarioDir(import.meta.url);
-
-beforeAll(async () => {
-  await installScenarioDependencies({ scenarioDir });
+const scenarioDir = await prepareScenarioDir({
+  scenarioDir: resolveScenarioDir(import.meta.url),
 });
+const autoHookScenarios = await getAISDKAutoHookScenarios(scenarioDir);
 
-for (const scenario of AI_SDK_AUTO_HOOK_SCENARIOS) {
+for (const scenario of autoHookScenarios) {
   test(`ai sdk auto-instrumentation via node hook collects ${scenario.agentSpanName} traces without manual wrapping (ai ${scenario.version})`, async () => {
     await withScenarioHarness(
       async ({ events, payloads, runNodeScenarioDir }) => {
