+++ e2e/scenarios/anthropic-auto-instrumentation-node-hook/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@braintrust/e2e-anthropic-auto-instrumentation-node-hook",
+  "private": true,
+  "braintrustScenario": {
+    "canary": {
+      "dependencies": {
+        "@anthropic-ai/sdk": "latest"
+      }
+    }
+  },
+  "dependencies": {
+    "@anthropic-ai/sdk": "0.60.0"
+  }
+}
