+++ e2e/scenarios/claude-agent-sdk-auto-instrumentation-node-hook/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "@braintrust/e2e-claude-agent-sdk-auto-instrumentation-node-hook",
+  "private": true,
+  "dependencies": {
+    "@anthropic-ai/claude-agent-sdk": "file:../../helpers/mock-claude-agent-sdk"
+  }
+}
