+++ e2e/scenarios/openai-auto-instrumentation-node-hook/scenario.openai-v4.mjs
@@ -7,4 +7,5 @@ import {
 runOpenAIAutoInstrumentationNodeHookOrExit(
   OpenAI,
   await getInstalledPackageVersion(import.meta.url, "openai-v4"),
+  "beta",
 );
